// topo-render/src/model/point/constants.ts
export enum PointType {
    /**
     * 
     */
    BOLT = 'bolt',
    /**
     * 
     */
    FEATURE = 'natural_protection',
    /**
     * 
     */
    ANCHOR = 'anchor',
    /**
     * 
     */
    GENERIC = 'generic',
}

export enum BoltType {
    /**
     * There are options here for rendering the thing that you like.
     */
    RING_BOLT = 'rb',
    /**
     * A 
     */
    U_BOLT = 'ub',
    /**
     * A 
     */
    FIXED_HANGER = 'fh',
    /**
     * When
     */
    CARROT = 'ct',
    /**
     * 
     */
    PITON = 'pn'
}